Refactor login to use Bloc pattern and add LoginView
Replaces AccountView with a new LoginView that uses the Bloc pattern for login state management. Adds LoginBloc, LoginEvent, and LoginState for handling login logic. Updates main.dart to use LoginView and provide LoginBloc. Removes the old AccountView implementation.
