+++ parkingapp/lib/blocs/login/login_event.dart
@@ -1,0 +1,15 @@
+import 'package:equatable/equatable.dart';
+
+abstract class LoginEvent extends Equatable {
+  const LoginEvent();
+  @override
+  List<Object?> get props => [];
+}
+
+class LoginSubmitted extends LoginEvent {
+  final String email;
+  final String password;
+  const LoginSubmitted({required this.email, required this.password});
+  @override
+  List<Object?> get props => [email, password];
+}
