+++ parkingapp/lib/blocs/login/login_state.dart
@@ -1,0 +1,20 @@
+import 'package:equatable/equatable.dart';
+
+abstract class LoginState extends Equatable {
+  const LoginState();
+  @override
+  List<Object?> get props => [];
+}
+
+class LoginInitial extends LoginState {}
+
+class LoginLoading extends LoginState {}
+
+class LoginSuccess extends LoginState {}
+
+class LoginFailure extends LoginState {
+  final String message;
+  const LoginFailure(this.message);
+  @override
+  List<Object?> get props => [message];
+}
